Update char-rnn-generation.ipynb
loss.data is now a 0-dim tensor it can't have elements. so loss.data[0] won't work. 
what he meant by loss.data[0] is a scalar of loss (a single value of loss). we can get the value by using loss.item() which throw out a single value of loss
so I replaced loss.data[0] with loss.item() and it works